feat: rename files in `shared` to match kebab-case
