+++ packages/shared/lib/hoc/with-error-boundary.tsx
